Room broadcast implemented
Websockets use redis pub/sub to broadcast messages across rooms.
Each room has a subscription channel.
A websocket server is subscribed to all rooms that it has connections for. On receiving a message from channel, it broadcasts the message to connections in that room.
Fixed some issues in the room creation/join process.